MAE model and more interesting samples
I don't think MAE was working correctly for the generators as it seems to get stuck at around a loss of 1...
The simulation log processing algorithm likely needs a complete rework after analyzing the MAESTRO dataset.
There are some interesting samples that sound quite nice at certain points but it's not ideal.
The discriminator eventually is able to accurately discriminate the samples far too easily with BCEwithlogitsloss as well so I likely wont be training any more due to the time we have left. Current setup however is useful for testing changes to the algorithm as we can easily track the losses after making changes.